Bugfixes and optimization 23
	Fix status bar text fields getting cut off in some cases
	Background: move all stack variables to top of function in MainWndProc(), and eliminate those unnecessary to reduce stack size
	Minor cleanup

+++ file_utils.c
@@ -387,6 +387,6 @@ BOOL DoSearch(LPCTSTR szText, LONG lStart, LONG lEnd, BOOL bDown, BOOL bWholeWor
 }
 #endif
 
-long ReplaceAll(LPTSTR* szBuf, long* bufLen, LPCTSTR szFind, LPCTSTR szRepl, BOOL bCase, BOOL bWholeWord){
+DWORD ReplaceAll(LPTSTR* szBuf, DWORD* bufLen, LPCTSTR szFind, LPCTSTR szRepl, BOOL bCase, BOOL bWholeWord){
 
 }
